+++ lisp/tree-sitter-tests.el
@@ -230,17 +230,18 @@ If RESET is non-nil, also do another full parse and check again."
 
 (ert-deftest minor-mode::node-at-point ()
   (tsc-test-lang-with-file 'rust "lisp/test-files/types.rs"
-    (should (eq 'source_file (tsc-node-type (tree-sitter-node-at-point 'source_file))))
+    (should (eq 'source_file (tsc-node-type (tree-sitter-node-at-pos 'source_file))))
+    (should (eq 'identifier (tsc-node-type (tree-sitter-node-at-pos nil 370))))
     (search-forward "erase_")
-    (should (eq 'identifier (tsc-node-type (tree-sitter-node-at-point))))
-    (should (eq 'function_item (tsc-node-type (tree-sitter-node-at-point 'function_item))))
-    (should (null (tree-sitter-node-at-point "function_item")))
-    (should (null (tree-sitter-node-at-point 'impl_item)))
+    (should (eq 'identifier (tsc-node-type (tree-sitter-node-at-pos))))
+    (should (eq 'function_item (tsc-node-type (tree-sitter-node-at-pos 'function_item))))
+    (should (null (tree-sitter-node-at-pos "function_item")))
+    (should (null (tree-sitter-node-at-pos 'impl_item)))
     ;; FIX: Signal an error for non-existing node types.
-    (should (null (tree-sitter-node-at-point 'non-existing-node-type)))
+    (should (null (tree-sitter-node-at-pos 'non-existing-node-type)))
     (search-forward "struc")
-    (should (equal "struct" (tsc-node-type (tree-sitter-node-at-point))))
-    (should (eq 'struct_item (tsc-node-type (tree-sitter-node-at-point 'struct_item))))))
+    (should (equal "struct" (tsc-node-type (tree-sitter-node-at-pos))))
+    (should (eq 'struct_item (tsc-node-type (tree-sitter-node-at-pos 'struct_item))))))
 
 (ert-deftest node::eq ()
   (tsc-test-with 'rust parser