+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [0.15.2] - 2021-09-12
 - Reduced GC pressure by not making the text property `face` a list if there is only one face.
 - Recast `tree-sitter-node-at-point` as more general `tree-sitter-node-at-pos`, taking optional POS argument.
 - Made `tree-sitter-node-at-pos` accept special node-type arguments `:named` and `:anonymous`.
@@ -122,7 +124,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.1.0] - 2020-01-27
 Initial release
 
-[Unreleased]: https://github.com/emacs-tree-sitter/elisp-tree-sitter/compare/0.15.1...HEAD
+[Unreleased]: https://github.com/emacs-tree-sitter/elisp-tree-sitter/compare/0.15.2...HEAD
+[0.15.2]: https://github.com/emacs-tree-sitter/elisp-tree-sitter/compare/0.15.1...0.15.2
 [0.15.1]: https://github.com/emacs-tree-sitter/elisp-tree-sitter/compare/0.15.0...0.15.1
 [0.15.0]: https://github.com/emacs-tree-sitter/elisp-tree-sitter/compare/0.14.0...0.15.0
 [0.14.0]: https://github.com/emacs-tree-sitter/elisp-tree-sitter/compare/0.13.1...0.14.0